requirements.txt
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
asttokens==2.0.5
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.6
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.3.9
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
codecov==2.0.15
coverage==4.5.3
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docopt==0.6.2
docutils==0.14
efficientnet-pytorch==0.7.1
einops==0.3.2
entrypoints==0.3
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.10
flake8==3.7.7
Flask==1.0.2
fonttools==4.33.3
future==0.17.1
glob2==0.6
hypothesis==4.23.2
idna==2.8
imageio==2.12.0
imagesize==1.1.0
inflect==2.1.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==1.1.0
jsonschema==3.0.1
jupyter-client>=5.2.4
jupyter-core>=4.4.0
jupyterlab==0.35.6
jupyterlab-pygments==0.2.2
jupyterlab-server==0.2.0
kiwisolver==1.1.0
libarchive-c==2.8
librosa==0.6.3
linformer==0.2.1
llvmlite==0.38.1
lmdb==0.94
MarkupSafe==1.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==2.0.0
more-itertools==7.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbclient==0.6.4
nbconvert==5.5.0
nbformat>=4.4.0
nest-asyncio==1.5.5
networkx==2.0
nibabel==3.2.1
nltk==3.4.1
notebook==5.7.8
numba==0.43.1
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.11.0
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
pure-eval==0.2.2
py==1.8.0
pybind11==2.2.4
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
pykwalify==1.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyradiomics==3.0.1
pyrsistent==0.15.1
PySocks==1.6.8
pytest==4.4.2
pytest-cov==2.7.1
pytest-pythonpath==0.7.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==3.0.2
pytz==2022.2.1
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
regex==2018.1.10
requests==2.21.0
resampy==0.2.1
revtok==0.0.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacrebleu==1.2.10
sacremoses==0.0.19
scikit-image==0.17.2
scikit-learn==1.1.2
scipy==1.8.1
Send2Trash==1.5.0
SimpleITK==2.2.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.8
spacy==2.0.16
Sphinx==2.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
stack-data==0.3.0
tabulate==0.8.3
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thinc==6.12.1
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.4.12
tinycss2==1.1.1
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets>=4.3.2
typing==3.6.6
typing-extensions==3.7.2
ujson==1.35
Unidecode==1.0.23
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2